+++ src/sources/index.ts
@@ -1,2 +1,3 @@
 export { transportDataGouvFr } from './transport-data-gouv-fr';
-export { mobilityData } from './mobility-data';
+export { mobilityData, createMobilityDataSource } from './mobility-data';
+export { mobilityDataCsv } from './mobility-data-csv';